+++ changelog.md
@@ -5,13 +5,13 @@
 - The shutdown event handlers will be ensured when a client connection receives a fatal exception.
 - Add `run` command.
 - Add `ApplicationFlag.application_command_badge`.
-- Add `delete_message_seconds` parameter to `Client.guild_ban_add`.
+- Add `delete_message_duration` parameter to `Client.guild_ban_add`.
 - Add `AutoModerationRuleTriggerType.mention_spam`.
 - Add `MentionSpamTriggerMetadata`.
 - Add `AutoModerationRule` parameter to `AutoModerationRule.__new__`.
 - Add `AutoModerationRule` parameter to `AutoModerationRule.copy_with`.
 - Separate down the `guild ban` endpoints from the `guild` ones.
-- Add `delete_message_seconds` audit log detail converter and update the `delete_message_days` one.
+- Add `delete_message_duration` audit log detail converter and update the `delete_message_days` one.
 - `Client.user_guild_profile_edit` now accepts `roles` as any iterable.
 
 #### Bug Fies
@@ -20,7 +20,7 @@
 
 #### Renames, Deprecation & Removals
 
-- Deprecate `delete_message_days` parameter of `Client.guild_ban_add`. Use `delete_message_seconds` instead.
+- Deprecate `delete_message_days` parameter of `Client.guild_ban_add`. Use `delete_message_duration` instead.
 
 ## 1.3.0 *\[2022-07-19\]*